+++ Sources/SwiftNotify/SwiftNotify.h
@@ -1,0 +1,19 @@
+//
+//  SwiftNotify.h
+//  SwiftNotify
+//
+//  Created by Johnny Choi on 9/11/2016.
+//  Copyright © 2016 Johnny Choi@Co-Fire. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for SwiftNotify.
+FOUNDATION_EXPORT double SwiftNotifyVersionNumber;
+
+//! Project version string for SwiftNotify.
+FOUNDATION_EXPORT const unsigned char SwiftNotifyVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <SwiftNotify/PublicHeader.h>
+
+